Refactor JSON Highlighter into a stateless component
I have extracted the hand-written JSON highlighter in `ModelDetail.tsx` into its own stateless helper component.

I moved the component from `orw-deno/islands/ModelDetail.tsx` to a new file at `orw-deno/components/JsonHighlighter.tsx`.

The new component is a pure, stateless function that does not rely on `useState` or `useEffect`, making it more efficient and reusable. This avoids the creation of an unnecessary Fresh island. I have also updated `ModelDetail.tsx` to use this new component.
